+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 * Make non-robot files visible in the Xill IDE project pane [CTC-1652]
 * Make non-robot files editable in Xill IDE [CTC-1617]
 * File.move construct [CTC-1643]
+* Validate Xill robot when uploading to Xill Server [XSVR-21]
 
 ### Change